Support Contexts
Add a second CallContext function that you can use to pass a
context.Context to an RPC call. This makes it easy to coordinate
cancelation across a number of different goroutines, and have one
failure make every request fail.

Add an example of RPC usage, and a test for the new context code.